exe may now be given to install-game cli
